HAN-35: move routes to main
