+++ app/main.py
@@ -1,7 +1,11 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, Request, status, Query
 from app.database.database import SQLAlchemyClient
 import logging
-from app.router import api_router
+from app.controller import example_controller
+from typing import List
+from app.schemas.example import (
+    ExampleSchema,
+)
 
 app = FastAPI()
 
@@ -31,4 +35,26 @@ async def shutdown_db_client():
 async def root():
     return {"message": "Hello World"}
 
-app.include_router(api_router)
+
+@app.post(
+    "/example",
+    status_code=status.HTTP_201_CREATED,
+    response_model=ExampleSchema
+)
+async def create_example(req: Request,
+                         item: ExampleSchema):
+    return example_controller.create_example(req, item)
+
+
+@app.get(
+    "/example",
+    status_code=status.HTTP_200_OK,
+    response_model=List[ExampleSchema]
+)
+async def get_example(req: Request,
+                      id_example: str = Query(None),
+                      limit: int = Query(10)):
+    if id_example is None:
+        return example_controller.get_all_example(req, limit)
+    return [example_controller.get_example(req, id_example)]
+
